refactor out zkos_prover lib
